pyproject.toml
[tool.poetry]
name = "django_telegram"
version = "0.1.0"
description = ""
authors = ["InzGIBA <a.baktubayev@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0.4"
pyjwt = "^1.7.1"
pyTelegramBotAPI = "^3.6.7"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
